Add templated message formatting
Includes a new utility function 'tsprintf()', which takes a string much
like the printf family, and a hash or hashref whose keys are variable
names that will be expanded to its value in the template string.
Variables in the template string are prefixed with '%' with optional
surrounding braces. Variables in the template string that are not
defined in the hash(ref) are ignored and returned as a literal part of
the string.