Switch from YAML configs to INI
I really love YAML, but the fact that there is seemingly only ONE YAML
parser in the wild that preserves comments—let alone key order—and is
for Python is very saddening. It is important that user-editable config
files retain their comments and order of settings, and no Perl YAML
implementation provides this.

ZeroBot::Config::File::hash and company is a bit of a hack, but will not
be present in ZeroBot v1.0; instead it will be replaced by a custom
config file format including parser and grammar.