Slight redesign of message types
Since messages, notices and actions can all in a sense be "private",
i.e. sent directly to a user, it makes more sense and is more useful to
make "privateness" an attribute, and not a specific message type.

MSGTYPE_PRIVATE has been removed in favor of adding a 'private'
attribute to the IRC::Message and IRC::Command classes. In addition,
there is no longer a separate Bot event for public and private messages;
there is now only Bot_irc_msg.

Lastly, respond() and emote() have been fixed and set the proper target.