Keep track of the currently connected server
IRC::Network now has a 'connected_server' attribute that holds
a reference to the IRC::Server object that the bot is connected to. In
addition, the 'get_server()' method will return the IRC::Server object
whose 'servername' matches the given argument.

IRC::Server now has a 'servername' attribute that holds the name of the
server *as reported by said server*. This distinction is made because
a misconfigured server can report a hostname that is different than the
one resolving to its address. E.g. the IRC server at example.com reports
that it is irc.example.com.

'servername' is initialized to the value of 'hostname' and changed
automatically if necessary, so 'servername' is the preferred attribute
unless you know you need one or the other.