Refactor IRC::Event and friends
IRC::Event is now a base class for "user-generated" events. Being
abstract enough for all IRC events was proving to be unnecessary and
restrictive. IRC::Answerable was consolidated into IRC::Event.

There are now events for nick changes, joins and kicks (more to come).
All original subclasses of IRC::Event now live in the IRC/Event
directory.

Modules can now use the respond() suite of methods on all events that
inherit from IRC::Event; it is up to the subclass to define exactly how
a reply() is different from a respond(), if applicable.